Simplify Makefile and allow for better incremental builds. (#3)
* Simplify Makefile and allow for better incremental builds.

Also add -std=gnu11 option since some older compilers default
to c99 which causes compilation to fail.

* Enable more warnings and fix some minor issues.

* Fixes a memory leak

Caused by method_name not being freed if
OQS_RAND_urandom_chacha20_ctx_new() fails. Also accounts for the fact
that strdup can fail and return NULL.

+++ src/rand/test_rand.c
@@ -20,7 +20,7 @@ static void rand_test_distribution_32(OQS_RAND *rand, unsigned long occurrences[
 	for (int i = 0; i < iterations; i++) {
 		x = OQS_RAND_32(rand);
 		uint8_t b;
-		for (int j = 0; j < sizeof(uint32_t); j++) {
+		for (size_t j = 0; j < sizeof(uint32_t); j++) {
 			b = (x >> j) & 0xFF;
 			OQS_RAND_test_record_occurrence(b, occurrences);
 		}
@@ -32,7 +32,7 @@ static void rand_test_distribution_64(OQS_RAND *rand, unsigned long occurrences[
 	for (int i = 0; i < iterations; i++) {
 		x = OQS_RAND_64(rand);
 		uint8_t b;
-		for (int j = 0; j < sizeof(uint64_t); j++) {
+		for (size_t j = 0; j < sizeof(uint64_t); j++) {
 			b = (x >> j) & 0xFF;
 			OQS_RAND_test_record_occurrence(b, occurrences);
 		}
